Move authorization create to folder
(cherry picked from commit 41e3e56eb0d721d080b8ce8d96a3e2b83a4aff34)
