+++ packages/graphql/tests/integration/mutations/create/authorization/create-authorization-filter.int.test.ts
@@ -17,8 +17,8 @@
  * limitations under the License.
  */
 
-import type { UniqueType } from "../../../utils/graphql-types";
-import { TestHelper } from "../../../utils/tests-helper";
+import type { UniqueType } from "../../../../utils/graphql-types";
+import { TestHelper } from "../../../../utils/tests-helper";
 
 describe("create with authorization", () => {
     const testHelper = new TestHelper();
